first stage finish

+++ routes/web.php
@@ -4,10 +4,14 @@ use App\Http\Controllers\AsistenciaController;
 use App\Http\Controllers\MiembroController;
 use App\Http\Controllers\MinisterioController;
 use App\Http\Controllers\PanelController;
+use App\Http\Controllers\PermissionController;
 use App\Http\Controllers\pruebaController;
+use App\Http\Controllers\RoleController;
+use App\Http\Controllers\UserController;
 use App\Http\Controllers\UsuariojcController;
 use App\Http\Controllers\UsuariosJcController;
 use Illuminate\Support\Facades\Route;
+use Illuminate\Validation\Rules\Can;
 
 /*
 |--------------------------------------------------------------------------
@@ -20,18 +24,11 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-/* Route::get('/', function () {
-    return view('index');
-})->middleware("auth"); */
-//restrinjo ruta register
-
-
-//Auth::routes();
 
 Auth::routes(["register" => true]);
 
 
-Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home')->middleware("auth");;
 
 
 //RUTA Principal 
@@ -46,63 +43,66 @@ Route::get('/', [PanelController::class, "index"])->middleware("auth");
 Route::get('/miembros', [MiembroController::class, "index"])->name("miembros")->middleware("auth");
 
 //Create Miembro
-Route::get('/miembros/create', [MiembroController::class, "create"])->name("miembros.create");
+Route::get('/miembros/create', [MiembroController::class, "create"])->name("miembros.create")->middleware("auth")->middleware("can:Gestionar Miembros");
 
 //Get Miembro
-Route::get('/miembros/show/{id}', [MiembroController::class, "show"])->name("miembros.show");
+Route::get('/miembros/show/{id}', [MiembroController::class, "show"])->name("miembros.show")->middleware("auth");;
 
 // Edit Miembro 
-Route::get('/miembros/edit/{id}', [MiembroController::class, "edit"])->name("miembros.edit");
+Route::get('/miembros/edit/{id}', [MiembroController::class, "edit"])->name("miembros.edit")->middleware("auth")->middleware("can:Gestionar Miembros");
 
 
 //Route procesa informacion 
 
-Route::post('/miembros/create', [MiembroController::class, "store"])->name("miembros.store");
+Route::post('/miembros/create', [MiembroController::class, "store"])->name("miembros.store")->middleware("auth");;
 
-Route::put('/miembros/update/{id}', [MiembroController::class, "update"])->name("miembros.update");
+Route::put('/miembros/update/{id}', [MiembroController::class, "update"])->name("miembros.update")->middleware("auth");;
 
-Route::delete('/miembros/delete/{id}', [MiembroController::class, "destroy"])->name("miembros.destroy");
+Route::delete('/miembros/delete/{id}', [MiembroController::class, "destroy"])->name("miembros.destroy")->middleware("auth");;
 
 
 //Route boton activo | inactivo 
-Route::get('/miembros/status/{id}', [MiembroController::class, "updateStatus"])->name("miembros.updateStatus");
-Route::get('/ministerios/status/{id}', [MinisterioController::class, "updateStatus"])->name("ministerios.updateStatus");
+Route::get('/miembros/status/{id}', [MiembroController::class, "updateStatus"])->name("miembros.updateStatus")->middleware("auth");;
+Route::get('/ministerios/status/{id}', [MinisterioController::class, "updateStatus"])->name("ministerios.updateStatus")->middleware("auth");;
 
 //END RUTAS MIEMBROS 
 
 
 //RUTAS MINISTERIOS 
 
-Route::resource("ministerios", MinisterioController::class);
+Route::resource("ministerios", MinisterioController::class)->middleware("auth");;
 
 //END RUTAS MINISTERIOS 
 
 
 //RUTAS USUARIOS 
 
-
+Route::resource("usuarios", UserController::class)->middleware("auth");;
 
 //END RUTAS USUARIOS 
 
 
 //RUTAS ASISTENCIA 
 
-Route::resource("asistencias", AsistenciaController::class);
+//vista reporte
+Route::get("asistencias/reporte", [AsistenciaController::class, "reporte"])->name("asistencias.reporte")->middleware("auth");;
+//una sola fecha
+Route::post("asistencias/reportePdf", [AsistenciaController::class, "reportePdf"])->name("asistencias.pdf")->middleware("auth");;
+//entre fechas
+Route::post("asistencias/reportePdf/fechas", [AsistenciaController::class, "reportePdfFechas"])->name("asistenciasFechas.pdf")->middleware("auth");;
+//rutas crud
+Route::resource("asistencias", AsistenciaController::class)->middleware("auth");;
 
 //END RUTAS ASISTENCIA
 
+//RUTAS Roles 
 
+Route::resource('roles', RoleController::class)->middleware("auth");;
 
+//END RUTAS Roles 
 
+//RUTAS Permissions 
 
+Route::resource('permissions', PermissionController::class)->middleware("auth");;
 
-
-
-
-//Pruebas 
-
-Route::get('/pruebagit', pruebaController::class)->name("pruebagit");
-
-Route::get('/prueba', function () {
-    return view("prueba");
-});
+//END RUTAS Permissions
